currency/requirements.txt
Django==2.2.7
celery==4.4.0
django-grappelli==2.13.4
djangorestframework==3.14.0
psycopg2-binary==2.9.7
redis==3.3.6
requests==2.31.0
xmltodict==0.13.0
importlib-metadata<5.0
